MyGPT/backend/requirements.txt
fastapi
uvicorn
openai
websockets
python-dotenv  # ✅ Add this line
